feat: Have search component actually do search with BooksAPI
Update the Search component to use the BooksAPI to look for search results in the server as the user is typing in results. Store results in state. Use moveBook method from App to add books to desired shelf. Intercept moveBook with addBook to navigate back to the users shelves when a book is added.